pyproject.toml
[tool.poetry]
name = "meteo-station"
version = "0.1.0"
description = ""
authors = ["Aleksey Kondratenko <kond@omegafuture.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
unicorn = "^2.0.1.post1"
pydantic = "^2.5.1"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
